endret navn fra spec til hendelseikon
